refactor(window): separate backend platform services
Separate window backend operations from platform services across SDL, Native,
Darwin, Linux, Android, and Windows implementations.

Centralize SIP constants, make safe-area no-ops explicit, and move resize
ownership to the selected backend so SDL screen changes do not depend on
native window code.
