+++ TotalCrossVM/src/nm/ui/Window.h
@@ -8,15 +8,8 @@
 #define WINDOW_H
 
 #include "tcvm.h"
+#include "WindowSIP.h"
 
-enum TCSIP
-{
-   SIP_HIDE    = 10000, /** Used to hide the virtual keyboard */
-   SIP_TOP     = 10001, /** Used to place the keyboard on top of screen */
-   SIP_BOTTOM  = 10002, /** Used to place the keyboard on bottom of screen */
-   SIP_SHOW    = 10003,  /** Used to show the virtual keyboard, without changing the position */
-   SIP_ENABLE_NUMERICPAD = 10004,
-   SIP_DISABLE_NUMERICPAD = 10005
-};
+bool windowBackendSetSize(int32 width, int32 height);
 
 #endif
