+++ TotalCrossVM/src/nm/ui/WindowSIP.h
@@ -1,0 +1,19 @@
+// Copyright (C) 2020-2021 TotalCross Global Mobile Platform Ltda.
+// Copyright (C) 2022-2026 Amalgam Solucoes em TI Ltda
+//
+// SPDX-License-Identifier: LGPL-2.1-only
+
+#ifndef WINDOW_SIP_H
+#define WINDOW_SIP_H
+
+enum TCSIP
+{
+   SIP_HIDE    = 10000, /** Used to hide the virtual keyboard */
+   SIP_TOP     = 10001, /** Used to place the keyboard on top of screen */
+   SIP_BOTTOM  = 10002, /** Used to place the keyboard on bottom of screen */
+   SIP_SHOW    = 10003,  /** Used to show the virtual keyboard, without changing the position */
+   SIP_ENABLE_NUMERICPAD = 10004,
+   SIP_DISABLE_NUMERICPAD = 10005
+};
+
+#endif
