+++ TotalCrossVM/src/nm/ui/linux/Window_c.h
@@ -1,32 +1,18 @@
 // Copyright (C) 2000-2013 SuperWaba Ltda.
-// Copyright (C) 2014-2020 TotalCross Global Mobile Platform Ltda.
+// Copyright (C) 2014-2021 TotalCross Global Mobile Platform Ltda.
+// Copyright (C) 2022-2026 Amalgam Solucoes em TI Ltda
 //
 // SPDX-License-Identifier: LGPL-2.1-only
-#if __APPLE__
-#include "SDL.h"
-#else
-#include "SDL2/SDL.h"
-#endif
 #include "../Window.h"
 
-static int32 sip;
-static bool windowGetSIP()
+static bool windowBackendSetSizeImpl(int32 width, int32 height)
 {
-   return sip != SIP_HIDE;
+   UNUSED(width)
+   UNUSED(height)
+   return false;
 }
 
-static void windowSetSIP(int32 sipOption)
-{
-   sip = sipOption;
-   if(sip == SIP_HIDE) {
-      SDL_StopTextInput();
-   }
-   else {
-      SDL_StartTextInput();
-   }
-}
-
-static void windowSetDeviceTitle(TCObject titleObj)
+static void windowBackendSetDeviceTitle(TCObject titleObj)
 {
    UNUSED(titleObj)
 }
